updating client app

+++ Tests/UnitTests/CntpGraphIrC/GraphIr/libb64/cdecode.c
@@ -40,7 +40,7 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
 				{
 					state_in->step = step_a;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (int)(plainchar - plaintext_out);
 				}
 				fragment = (char)base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -51,7 +51,7 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
 				{
 					state_in->step = step_b;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (int)(plainchar - plaintext_out);
 				}
 				fragment = (char)base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -63,7 +63,7 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
 				{
 					state_in->step = step_c;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (int)(plainchar - plaintext_out);
 				}
 				fragment = (char)base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -75,7 +75,7 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
 				{
 					state_in->step = step_d;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (int)(plainchar - plaintext_out);
 				}
 				fragment = (char)base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -83,6 +83,6 @@ int base64_decode_block(const char* code_in, const int length_in, char* plaintex
 		}
 	}
 	/* control should not reach here */
-	return plainchar - plaintext_out;
+	return (int)(plainchar - plaintext_out);
 }
